lima/ppir: change offset type to int
Offset doesn't need to be 64-bit. This fixes compilation error
with 64-bit off_t.

Fixes: af0de6b9 lima/ppir: implement discard and discard_if

Suggested-by: Qiang Yu <yuq825@gmail.com>
Tested-by: Andreas Baierl <ichgeh@imkreisrum.de>

+++ src/gallium/drivers/lima/ir/pp/ppir.h
@@ -297,7 +297,7 @@ typedef struct ppir_instr {
    int est; /* earliest start time */
    int parent_index;
    bool scheduled;
-   off_t offset;
+   int offset;
 } ppir_instr;
 
 typedef struct ppir_block {
